- Major upgrade: Python3 compatibility

+++ pyorbit/classes/io_subroutines.pyx
@@ -115,9 +115,9 @@ def nested_sampling_load_from_cpickle(output_directory, prefix=''):
 def emcee_flatchain(chain, nburnin, nthin):
     """flattening of the emcee chains with removal of burn-in"""
     _, d, _ = np.shape(chain)
-    nburn = nburnin / nthin
+    nburn = int(nburnin / nthin)
     if nburn >= d*0.9:
-        nburn = d/4
+        nburn = int(d/4)
 
     s = chain[:, nburn:, :].shape
     return chain[:, nburn:, :].reshape(s[0] * s[1], s[2])
@@ -127,18 +127,18 @@ def emcee_flatlnprob(lnprob, nburnin, nthin, emcee_version):
     if emcee_version == '3':
         """flattening of the emcee chains with removal of burn-in"""
         d, _ = np.shape(lnprob)
-        nburn = nburnin / nthin
+        nburn = int(nburnin / nthin)
         if nburn >= d * 0.9:
-            nburn = d / 4
+            nburn = int(d / 4)
 
         s = lnprob[nburn:, :].shape
         return lnprob[nburn:, :].reshape(s[0] * s[1])
     else:
         """flattening of the emcee chains with removal of burn-in"""
         _, d = np.shape(lnprob)
-        nburn = nburnin / nthin
+        nburn = int(nburnin / nthin)
         if nburn >= d * 0.9:
-            nburn = d / 4
+            nburn = int(d / 4)
 
         s = lnprob[:, nburn:].shape
         return lnprob[:, nburn:].reshape(s[0] * s[1])
